+++ test/t_load.cpp
@@ -4,7 +4,7 @@ using namespace std;
 
 int main() {
 	char path[] = "../data/layer_1_weight.txt";
-	double * weight = new double[2*20];
+	float * weight = new float[2*20];
 	load(path, weight);
 	for (int i = 0; i < 40; i++)
 		cout <<weight[i] << endl;
